tests fixes

+++ test-server-module/src/main_types/vectors.rs
@@ -15,10 +15,10 @@ pub struct Vector3 {
 
 #[derive(SpacetimeType, Debug, Clone, Copy)]
 pub struct Vector4 {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
-    pub w: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
+    pub w: f32,
 }
 
 impl Vector3 {
